add translation for welcome email
