+++ app/mailers/user_mailer.rb
@@ -5,7 +5,7 @@ class UserMailer < ApplicationMailer
   def welcome_email
     @user = params[:user]
     @url  = ENV.fetch('SERVER_HOSTNAME')
-    mail(to: @user.email, subject: 'Your Nurture account is now registered!')
+    mail(to: @user.email, subject: I18n.t('user_mailer.welcome_email.text_subject'))
   end
 
   def update_email
